requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==21.2.0
pymongo==4.0.1
python-dotenv==1.0.1
pymongo[srv]==3.11